Move formatting into the crate, gated by a feature
